feat: Fix back to outline link
The "back to course outline" link was pointing to the index.html page which is password protected. This was causing the user to be prompted for a password every time they clicked the link.

This commit fixes the issue by pointing the link to the main.html page which is the course outline.